feat: Upload watchdog events #395
* renamed files to align with google functions deployments.
* started using google-cloud-logging library
