+++ bigquery/cloud-functions/diagnostics/requirements.txt
@@ -1,3 +1,3 @@
 google-cloud-storage==1.42.2
 google-cloud-bigquery==2.28.1
-hm-pyhelper==0.13.22
+google-cloud-logging==3.2.1
